+++ richat/src/channel.rs
@@ -12,7 +12,7 @@ use {
     },
     richat_proto::{
         geyser::{
-            CommitmentLevel as CommitmentLevelProto, SubscribeRequest,
+            CommitmentLevel as CommitmentLevelProto, SlotStatus, SubscribeRequest,
             SubscribeRequestFilterAccounts, SubscribeRequestFilterBlocksMeta,
             SubscribeRequestFilterEntry, SubscribeRequestFilterSlots,
             SubscribeRequestFilterTransactions,
@@ -206,7 +206,10 @@ impl Messages {
                     ConfigGrpcClientSource::DragonsMouth => connection
                         .subscribe_dragons_mouth_once(SubscribeRequest {
                             accounts: hashmap! { "".to_owned() => SubscribeRequestFilterAccounts::default() },
-                            slots: hashmap! { "".to_owned() => SubscribeRequestFilterSlots::default() },
+                            slots: hashmap! { "".to_owned() => SubscribeRequestFilterSlots {
+                                filter_by_commitment: Some(false),
+                                interslot_updates: Some(true),
+                            } },
                             transactions: hashmap! { "".to_owned() => SubscribeRequestFilterTransactions::default() },
                             transactions_status: HashMap::new(),
                             blocks: HashMap::new(),
@@ -269,9 +272,9 @@ impl Sender {
                     self.confirmed.push(slot, message.clone());
                     self.finalized.push(slot, message.clone());
 
-                    if let Some(sender_shared) = match msg.commitment() {
-                        CommitmentLevelProto::Confirmed => Some(&mut self.confirmed),
-                        CommitmentLevelProto::Finalized => Some(&mut self.finalized),
+                    if let Some(sender_shared) = match msg.status() {
+                        SlotStatus::SlotConfirmed => Some(&mut self.confirmed),
+                        SlotStatus::SlotFinalized => Some(&mut self.finalized),
                         _ => None,
                     } {
                         if let Some(slot_info) = self.slots.get(&slot) {
@@ -283,7 +286,7 @@ impl Sender {
                     }
 
                     // remove slot info
-                    if msg.commitment() == CommitmentLevelProto::Finalized {
+                    if msg.status() == SlotStatus::SlotFinalized {
                         loop {
                             match self.slots.keys().next().copied() {
                                 Some(slot_min) if slot_min <= slot => {
@@ -593,8 +596,8 @@ impl SlotInfo {
         // mark as landed
         if let ParsedMessage::Slot(message) = message {
             if matches!(
-                message.commitment(),
-                CommitmentLevelProto::Confirmed | CommitmentLevelProto::Finalized
+                message.status(),
+                SlotStatus::SlotConfirmed | SlotStatus::SlotFinalized
             ) {
                 self.landed = true;
             }